pyproject.toml
[project]
name = "accessnine-shoppingai-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "김동환", email = "kimdonghwandeveloper@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[project.scripts]
accessnine-shoppingai-rag = "accessnine_shoppingai_rag:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
